+++ core/src/run_steps.ts
@@ -150,22 +150,4 @@ function cancelRun(logger: LoggerInterface, collections: Collection[]) {
   });
 }
 
-class Jackrabbit {
-  #logger: LoggerInterface;
-  #collections: Collection[];
-
-  constructor(logger: LoggerInterface, collections: Collection[]) {
-    this.#logger = logger;
-    this.#collections = collections;
-  }
-
-  start() {
-    startRun(this.#logger, this.#collections);
-  }
-
-  stop() {
-    cancelRun(this.#logger, this.#collections);
-  }
-}
-
-export { Jackrabbit };
+export { startRun, cancelRun };
